payroll/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
django-filter
mysqlclient
